gollvm: fix variable context for aggregate types
CL 44014 updates the load/store mechanism. But variable context
determination is not updated: it still returns lvalue for
aggregate types, even it is small. As a consequence, for assignment
of a small struct type, it generates a pointer to RHS, whereas
the LHS expects a value, which crashes the compiler.

This CL changes variable context determination to use the new
useCopyForLoadStore. As it depends on useCopyForLoadStore, move
varContextDisp to Llvm_backend class (and it is only called in
this class).

Doing this, I get into some problems:
- it starts to generate load for 0-sized types (probably harmless).
- some trouble with function call argument passing, for example,
  for f([2]float32), the C-ABI oracle decides the parameter type
  to be <2 x float>, whereas the caller passes in [2 x float]...
  I have not yet resolved this problem.

For the moment, it may be easier to always use llvm.memcpy for
aggregate types. The backend seems to be smart enough to rewrite
it to load/stores for small types. We can revisit it if this
turns out to be a problem.

Undo the change on BackendPointerExprTests.cpp in CL 44014 (back
to memcpy).

Change-Id: Id1eac1242d18079d0496982d81fa3f05b08e1ac9
Reviewed-on: https://go-review.googlesource.com/44172
Reviewed-by: Than McIntosh <thanm@google.com>