gollvm: remove overly strict assert in conditional_expression
Remove an assert in Llvm_backend::conditional_expression that was
trying to enforce exact type matching between the "then" and "else"
expressions. Exact match is not always possible, since we can have
things like an untyped nil pointer expression on one arm and some
other concrete pointer type on the other arm.  Leaving the assert out
in this case should be fine, since if there is a real type clash, it
will be caught in the temp var assignment.

Change-Id: Id840ecd4e92eb077370fdfcbb849747c55f4df98
Reviewed-on: https://go-review.googlesource.com/44272
Reviewed-by: Ian Lance Taylor <iant@golang.org>