gollvm: revamp handling of complex types
Although LLVM has no notion of a complex type (all complex values and
operations have to be lowered beforehand), there needs to be an
awareness of complex values/variables for debug generation.  To help
with this create a new Btype subclass type for complex and update a
few other places in the bridge to handle the new subtype. Change debug
info meta-data generation for complex variables to emit the correct
DWARF tag. Note that this change doesn't include actual support for
doing anything with complex values/variables yet.

Change-Id: Ic000f1e7702f5ca76e0e19174802ceb88e5b3f07
Reviewed-on: https://go-review.googlesource.com/44433
Reviewed-by: Cherry Zhang <cherryyz@google.com>