gollvm: revamp temporary management
Rework the way that backend-generated temporary variables are created
and managed. Prior to this point the only interface for temp creation
was associated with the Bfunction class, meaning that in many places a
pointer to the Bfunction currently being constructed had to be passed
from one place to another in case a temp needed to be created (very
cumbersome). Instead of doing this, create temporary allocas() on the
fly and store them in a set in the node builder, then detect uses of
temps when assigning instructions to BBs (so as to rehost/reparent the
alloca instructions appropriately).

Change-Id: I9574aa03b553d6edd077d9f9169854d94370d91d
Reviewed-on: https://go-review.googlesource.com/44860
Reviewed-by: Cherry Zhang <cherryyz@google.com>