gollvm: more relaxed function-value typechecking
This change is designed to relax assignment-compatibility typechecking
to allow for more interchangeability between raw pointer-to-function
values and function descriptors (struct with single field composed of
pointer-to-function). There are some cases involving interface values
where the front end uses makes assignments between these two types
without a conversion, which we want to support.

Change-Id: Idf1b2642d1d04b56a99602afbb540a5a64e03577