gollvm: support struct_field_expression and array_index_expression on composite data
For a non-constant composite struct/array, its generation is
delayed until we can resolve its storage. To take an element
from it, we need to resolve it by storing into a temp variable,
then load from it.

For constant composite struct/array, either take its field
directly, or materialize it into a temp variable and load from
it if the index is not constant.

Change-Id: I1e4b4ed66d445277287be0c9f51ef38ba294367a
Reviewed-on: https://go-review.googlesource.com/44970
Reviewed-by: Than McIntosh <thanm@google.com>