gollvm: add type equivalence helper
Add a new Btype::equivalent method that supports comparing two Btype
objects for structural equivalence: equal LLVM types and Btype
properties (ex: integer signed-ness) but ignoring type names for the
purposes of the comparison. This is in preparation for a follow-on CL
to fix some bugs in Llvm_backend::convert_expression().

Change-Id: I4c6bd5f4ada3169c2e372414ad090869ac0755da