gollvm: rework some handling of constant
In some cases, an expression is not a constant even the
underlying value is an llvm::Constant. For example, for a
var expression of a global variable, its value is an
llvm::GlobalVariable, which is an llvm::Constant, but
generally we don't want to treat it as a constant
expression.

Introduce and use a wrapper function to determine whether
an expression has a constant value.

Change-Id: I49c19e50375cd3fa11e413bbb4a0be3010dc6674
Reviewed-on: https://go-review.googlesource.com/45932
Reviewed-by: Than McIntosh <thanm@google.com>