gollvm: fix call with constant complex argument
When passing composite constant as argument, we need to spill it
to memory so it can be cast to 8-byte objects (for example,
{float, float}* to [2 x float]*).

This is useful in passing complex constant. In theory this could
also happen for constant struct, but the FE inserts a temporary
varaible holding the constant value first.

Change-Id: Iff8bf4b07c573a94b106aedbdebbe030669e33c9