gollvm: use correct type size utility in CABI code
Change the CABI code to call TypeManager::typeSize() when requesting
the size for argument and return values, instead of dropping down to
directly call LLVM's typeAllocSize. The latter will assert if some
component of a type is unsized, which can happen if there is an
unresolved placeholder pointer type somewhere within it, whereas the
helper TypeManager::typeSize already has machinery to handle this
case.

Change-Id: Ib9702b60bac1580860675d2dcdb1a29b159dc13d