gollvm: revise recipe for skipping redundant conversions
Rework the machinery that skips creation of useless type conversions
in Llvm_backend::materializeConversion -- basing the decision on a
call to Btype::equivalent is not a good strategy. Instead, if llvm
types match but Btypes do not, create a conversion node (to record the
new Btype) but don't materialize any new llvm cast(s).

Change-Id: I4cefb1c720ac2c39508228ad666d70d59755348a