gollvm: fix buglet in CABI handling (constant arg spill)
Call argument marshalling has code that spills a composite constant
value to memory prior to copying them into proper CABI type needed at
the call. This code wasn't properly distingishing between a pure
anonymous constant value (ex: [2]float64{10.0, 10.0}) and a reference
to a global constant of the same type. In the latter case there is not
need to spill, since the value in question will be the address of the
global.

Change-Id: I8e74f10983a707a4d4010c6ad8e8ffaa46484ea3