gollvm: synthesize LHS/RHS context for var expressions
Up until now the bridge code would rely on the frontend to pass in a
tag on each var expression to indicate whether the variable appeared
in a left-hand-side or right-hand-side context.  This change adds
machinery to synthesize the correct tags from within the bridge,
without any input from the front end.

Also as part of this change:

- move the expansion of complex operations (compare, arithmetic, etc)
  a bit earlier in the compiler flow, so as to avoid a lot ofu
  materialize() / resolve calls along the way

- revamp handling of compound expressions (N_Compound) to be more
  consistent with other expressions with respect to llvm::Value:
  initial Bexpression will have a NULL llvm::Value, then a value is
  assigned during materialization.

Change-Id: Ic70e62d20122118bb123d7babb7ae64ef45a75e9
Reviewed-on: https://go-review.googlesource.com/47930
Reviewed-by: Cherry Zhang <cherryyz@google.com>