gollvm: support skipping specific subtrees in Bnode visitor
Add hooks to the templatized Bnode walk utility to allow for skipping
a specific subtree during the walk. This works by having the
visitChildPre callback return an additional flag that tells the walker
to avoid descending into the child subtree in question.  Use this
mechanism in FoldVisitor and MaterializeVisitor to avoid revisiting
expression trees that have already been evaluated (for example, the
Bstatement child of a compound expression).

Change-Id: I18cc8e53d9bc9f7ae4e4cab4938146e02a0f9ba0