gollvm: fix issue with integer constant creation
The recipe used for constructing integer constant expressions in the
bridge was not properly handling the case where you have a signed
mpz_val with value ((int64_t) -1) being forced into an unsigned
integer Btype with fewer bits (ex: uint16 or uint32). This triggered
an assert, since the mpz_val was being converted to a uint64 value of
18446744073709551615, then the llvm::ConstantInt::isValueValidForType
check for the smaller type failed. Fixed code converts constant
value into an llvm::APInt with the correct signed-ness and number of
bits before invoking llvm::ConstantInt::get.

Change-Id: I68741eea6833705cffe2eb06e33c3e771072aaeb