gollvm: tweak struct field GEP recipe to avoid type clash
Modify the IRBuilder call used to create structure field GEPs
to avoid passing in the struct type (which tells the builder
to infer the type from the storage pointer). This is a somewhat
hacky change, but helps in situations where there are LLVM typing
disagreements due to intermixing of function descriptor values
and raw pointer-to-function values.

Change-Id: I2f7588ab77adbf1d847a3f25ee180803376c521f