gollvm: friendlier failure mode for error compilations
Currently if a compilation fails due to syntax errors, it is not
uncommon for the driver to terminate in an assert while deleting the
LLVM context, due to dangling/orphaned llvm instructions.

To help terminate more gracefully on errors, enhance the BnodeBuilder
to do cleanup any dangling/orphaned instructions in its destructor,
and reorder things in the top-level driver so that the backend (and
with it the BnodeBuilder) is destroyed as soon as IR creation is
complete. This will also reduce the amount of memory needed for the
compile, since memory freed by BnodeBuilder will be available for LLVM
passes.

Change-Id: Iea2348219cc47ed6bca77579eb1833b480101fbb
Reviewed-on: https://go-review.googlesource.com/50651
Reviewed-by: Cherry Zhang <cherryyz@google.com>