gollvm: delay circular pointer address wraparound
Previously, the circular pointer type is wrapped around
when an address expression of such type is materialized.
For example, for

  type T *T
  var t T

when an address expression &t is made, the type is
wrapped around from *T to T. This may cause problem
when a *T type is indeed needed, for example, assigning
to a variable of type *T.

This CL delays the wraparound when the expression is
used and the type wraparound is indeed needed.

Change-Id: I0145e7974d4c6a651ae438d876f5c18c9e8118a8