gollvm: additional circular type workarounds
Put in place some additional workarounds for circular types during the
debug info meta-data build process. The scenario here is that we have
a function with type

   void (i8*, %B*)

where B is a structure that incorporates a function descriptor
whose type is the function above. In this case we discover the
cycle when processing the pointer-to-function type, which wasn't
handled before.

This is clearly not an ideal solution, but I want to make sure this
workaround is in place for the moment since this problem comes up
very frequently in Go test code.

Change-Id: Iaccf3e0aa1263ea21fe0c43da77a3cf08255344a
Reviewed-on: https://go-review.googlesource.com/50872
Reviewed-by: Cherry Zhang <cherryyz@google.com>