gollvm: fix function arg debug info generation
For function arg that "passed by reference", it may have no
initializer. For debug info generation its initializer is set
to the first instruction in the function entry block. However,
this may not work when the function entry block is empty or
contains only branch instruction. Fix this by generating the
declaration at function entry directly, instead of using a
dummy initializer.

Change-Id: I2b1438aa61e78fc7c5a08b19d389dde8444a090d