gollvm: fix buglet in handling of named types
If the front end creates a placeholder struct type, then creates
a named version of the placeholder (via Backend::named_type),
this can result in an assert when the last placeholder field
of the struct is updated (double call to setBody). Check for this
case and avoid the double update.

Change-Id: I082eeb61146fa663b3a0f72e1aa7005e857990c5