gollvm: fix buglet in TypeManager::fcnPointerCompatible
Fix a bug in the type manager utility routine for determining if two
LLVM types are function-pointer-compatible: redo the way cycles are
checked for. Specifically if we're comparing the following two types for
equivalence:

A:  { %_type.0*, i32 (i8*, i8*)*, i32 (i8*, i8*)*, i8* }
B:  { %_type.0*, i32 (i8*, i8*)*, %__go_descriptor.0*, i8* }

when considering the third field of the struct, the type in struct "A"
will already be part of the visited set (since it was encountered in
field 1) but the type in struct "B" will not be visited -- this clash
should not cause the check to fail. To resolve this, test for
function-pointer/function-descriptor equivalence before performing the
visited check, which will treat the two types above as equivalent.

Change-Id: Ieef2ddc9496ac596dcd08752f460f0fa1b01e0f1