Update patch set 2
Patch Set 2: Code-Review+2

The change looks good. However, in gofrontend/go/backend.h, for named_type it says
  // This will not be called for a type
  // created via placeholder_pointer_type, placeholder_struct_type, or
  // placeholder_array_type.
so there might be something odd going on if named_type is called on a placeholder struct type. On the other hand, I can see the double setBody scenario happens, for example, with a named type which is a struct_type that contains a placeholder field. I think this is what's going on in the CL.

Patch-set: 2
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2