gollvm: update expression's value after postProcessInst
The instruction corresponding to an expression's value may
be changed/replaced in postProcessInst. In this case, update
the expression's value to point to the new instruction.

Unfortunately Bexpression::setValue has to be exported.
Bexpression::setStoreValue is exported but not used anywhere,
removed.

Found this from a program like the following:

func f() {
	defer ...
	switch g() {
		...
	}
}

The defer statement makes the call instruction (call g()) be
rewritten to an invoke instruction. The call expression's
value is later used in genSwitch.

Change-Id: I1552b563aaa82a51c835f5886f517fdf39bea1d4
Reviewed-on: https://go-review.googlesource.com/51390
Reviewed-by: Than McIntosh <thanm@google.com>