gollvm: handle circular function type in isPtrToArrayOf
In the following code,

type CFT func(...) // circular function type
var x []CFT = { func ... }

When making the composite value for the slice x, it stores the
pointer to the array of function pointers (a value of type
[n x void(...)*]*) in the slot of pointer to circular function
type (CFT.0**). Handle this type conversion.

A bit hacky here in that we assume the type matches when the
array element types of both the from-type and the to-type are
circular function type, without checking they are indeed the
same function.

Change-Id: I6539680a28ad7e783c5cb52e7be8cf80dfdc5594
Reviewed-on: https://go-review.googlesource.com/51991
Reviewed-by: Than McIntosh <thanm@google.com>