gollvm: handle assignments with degenerate LHS exprs
The bridge code that materializes indirection/deref expressions was
getting confused in situations where you have a constant value being
cast to a pointer and then deferenced on the LHS of an assignment,
e.g. "*(*sometype)(unsafe.Pointer(uintptr(<constant>))) = ...",
presumably as a means of generating a fault/crash.  Detect such cases
and handle them correctly.

Change-Id: I171f6e6a9410938217a286ac2bf41ba1f431b493