gollvm: handle multiple declarations for same external fcn
The go std library's syscall package creates two declarations for
"syscall", one for 64-bit calls and the other for 32-bit calls, giving
each the same external name. This doesn't play well with LLVM, since
within a given llvm::Module one can only have a single symbol with a
given name (aliases are allowed, but aliased symbols must share the
same type). Resolve this conflict by checking during Bfunction
creation to see if an existing symbol exists for the name being
declared; if so, create the second Bfunction with a value returned
from Module::getOrInsertFunction, which will be a bitcast of the
existing Fcn with the new type.

Change-Id: I3f4b3965141bf0a24c626369e3aab9a2da1a3762
Reviewed-on: https://go-review.googlesource.com/55190
Reviewed-by: Cherry Zhang <cherryyz@google.com>