Update patch set 2
Patch Set 2:

I was thinking about this a little bit last week (but didn't do much). I think there may be more weird but valid cases, like:

        p := uintptr(...)
        *(*int32)(unsafe.Pointer(p+8)) = 42

Here, p+8 is not a constant, but also not a pending var expr. This CL handles only constants, which seems too narrow.

Also, a weirder one:
        p := float64(0x1234)
        *(*int32)(unsafe.Pointer(uintptr(int64(p)))) = 42;

This currently compiles, but incorrectly:
  %cast.0 = bitcast double* %p to i64*, !dbg !13
  %cast.1 = bitcast i64* %cast.0 to i8**, !dbg !14
  %cast.2 = bitcast i8** %cast.1 to i32**, !dbg !15
  %deref.ld.0 = load i32*, i32** %cast.2, !dbg !16
  store i32 42, i32* %deref.ld.0, !dbg !17

This should be compiled to a store to address 0x1234, instead of a store to address 4661845738886529024 (the bit representation of float64 0x1234). The pointer cast from double* to i64* is problematic.

I think it probably makes more sense to recognize the expr being dereferenced as rvalue, even if the deref expression itself is an lvalue. So, for the float64 case above, it should do a load for float64 first, then convert to int64. I haven't tried this though. There may be cases I missed.

Let me know what you think and if I should try it. Alternatively we could get this CL in first, so at least constant case is handled.

Patch-set: 2
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>