Update patch set 2
Patch Set 2:

Regarding this one:


Regarding this one:

        p := uintptr(...)  *(*int32)(unsafe.Pointer(p+8)) = 42

I'll post a revised patch that should work for this.

For the weirder case, e.g.

        p := float64(0x1234)
        *(*int32)(unsafe.Pointer(uintptr(int64(p)))) = 42;

I agree that the IR we get for this is definitely not right.

I think it boils down to how we want to interpret casts on the LHS,
e.g. for

   var x float64 ... op(int64(var(x))) ... =

At the moment this is interpreted as "take &x [of type float64*] and
cast it to type int64*, then do some operation "op" on it.  I am
pretty sure that the IR created by gofrontend relies on this, although
now that I write this I think maybe I should double-check.

Patch-set: 2
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>