gollvm: don't propagate lvalue-ness down on deref expression
For *expr = val, expr should be evaluated as an rvalue, even
the deref expression (*expr) is an lvalue. For example,

	var p float64 = ...
	*(*int32)(unsafe.Pointer(uintptr(p))) = 5

The expression being dereferenced should be evalueated as a
float64 load of p, a convertion to uint then to *int32,
instead of a pointer bitcast from *float64 to *uint then load.

Change-Id: Ife46a4f6181636e9561c7968891cfac67421743f