gollvm: fix exception handling stmt bugs
Revise Llvm_backend::exception_handler_statement(), which was not
correct as implemented. Specifically, if an exception is thrown by
something in the catch clause, the catch block for it needs to store
the exception and make a note that the exception took place... then
after the "finally" block has executed, the cleanup code should
rethrow the exception (if there was one) as opposed to unconditionally
returning.

Change-Id: Ia63297cad1bc601f9ab5066ad879b1bc5739efd1