gollvm: check for no-pred blocks in unit test harness
Enhance the unit testing harness to detect and complain about
blocks with no predecessors.  Although these are sometimes harmless,
they can sometimes also be an indication of a bug, so it's good to
have a way to find out about them. Fix a couple of places in the
control flow generation code to avoid creating orphan blocks.

Change-Id: I2adb3b07952a528429e67a178ba95cb62571bacf