gollvm: handle function declaration and definition with different types
For some runtime functions called by the compiler, the compiler
may generate calls with dummy types, for example, the compiler
turn "make(chan ...)" to a call to runtime.makechan with dummy
pointer type, which in turn generates a declaration as

func makechan(t *_type, size int64) *struct{}

When compiling runtime, however, there is the actual definition
with the actual type,

func makechan(t *chantype, size int64) *hchan

The existence of both causes problems.

This CL handles this kind of type mismatch. In this case, the
type of the definition wins, and the declaration with the other
type is turned into a bitcast.

Change-Id: I987137d2bfcae05778c377677a35a3768fa5f0eb