gollvm: fix function declaration and definition with same name
This is similar to CL 57790, which handles function declaration
and definition with same name and different types. But it could
also happen that they have the same type (in fact this is the
common case). That CL doesn't cover this situation. This CL
fixes it.

There seems also similar thing to be done with global variables.
Will look into it and do it in a later CL.

Change-Id: I4ff99c35804d3788d3b34956410275c81ef7699f
Reviewed-on: https://go-review.googlesource.com/58970
Reviewed-by: Than McIntosh <thanm@google.com>