gollvm: fix global variable declaration and definition with same name
Along the same line as CL 55190, CL 57790, CL 58072, and
CL 58970, for global variables.

For certain runtime variables directly used by the compiler
(e.g. runtime.writeBarrier), the compiler creates a declaration
of external variable. When compiling the runtime, it also emits
the actual definition. They both must refer to the same
underlying global variable. Otherwise we end up creating two
versions of the global var, one used in some functions while
the other used in some other functions, and things fall apart.
This CL handles this case. In particular, the types may
mismatch: the compiler emits declaration of
runtime.writeBarrier as int32 (because it only needs to load
the value and test it against zero), but the actual definition
in the runtime is a struct.

Change-Id: I09ce55282de47337d4ef66678bab105d8c34f0a1
Reviewed-on: https://go-review.googlesource.com/59052
Reviewed-by: Than McIntosh <thanm@google.com>