gollvm: fast path for creating zero valued composite constant
Detect if we're creating a zero value, and if so take a fast
path. This makes "[1<<30]byte{}" not take forever.

Nonzero value like "[1<<30]byte{1<<29: 42}" is still very slow.
gccgo just sets the element inlined without creating a gigantic
literal in static data. Maybe we can do this as well. This will
rely on a fast creation of zero value. So this CL sets up the
base.

Change-Id: I15a9cf828e528a99c3cd72be1fb03344156687c8
Reviewed-on: https://go-review.googlesource.com/63271
Reviewed-by: Than McIntosh <thanm@google.com>