gollvm: support one invocation with .go and .o inputs in the wrapper
gccgo's testing framework may invoke gccgo with both .go files
and .o files as input, which does compilation and linking in
one run. Add this support in the wrapper. If the wrapper is
invoked with both .go and .o inputs, it skips the .o files in
compilation, but records them and passes them to the linker.
This makes "rundir" action in gccgo's go.test work.

Change-Id: I3446f0f96e34b93c738075ea49c02d3062a031b3
Reviewed-on: https://go-review.googlesource.com/63590
Reviewed-by: Than McIntosh <thanm@google.com>