gollvm: fix compiler assert on duplicate parameter
When the front end encounters a syntax error of some sort with a
parameter in a function definition (ex: func f(x int, x bool)) it will
continue the compilation, but won't go on to create the Bvariable for
the parameter. This can confuse things later on, in case that expects
to see entries for all of the params.  Check for this case in
Llvm_backend::function_set_parameters and set a flag on the Bfunction
in question, so that later on we can work around the missing param
(and avoid any asserts/crashes).

Change-Id: I51cd9aee9f5ec10270d6a932f471bc0889612ccc
Reviewed-on: https://go-review.googlesource.com/63671
Reviewed-by: Cherry Zhang <cherryyz@google.com>