gollvm: type fix in convertForAssignment
In convertForAssignment, we are converting the LLVM value srcVal
to the right type, so it makes more sense to use srcVal's type
instead of srcBType.

One case where this makes a difference is that the value is a
pointer to function descriptor, and the btype is another pointer
to function descriptor type which is a placeholder pointer type
redirected to the value's type. This appears in building the
testing package.

Change-Id: I849e98cb7c89a052ab990eba2a76c966cddf8894
Reviewed-on: https://go-review.googlesource.com/64010
Reviewed-by: Than McIntosh <thanm@google.com>