gollvm: bug fixes for globals with zero sized type.
Revise the way the bridge handles zero-sized globals, to account for
unpleasant behavior in Linux linkers (ld.bfd, ld.gold) related to
zero-sized things.

The linker bugs/problems in question crop up not with "normal" Go
programs (in which all of the interesting Go code exists within a
single load module) but in cases where you have a chunk of Go code in
one module being linked against a shared library containing other bits
of Go code (the canonical example being the shared version of
libgo). Scenario:

     abc.go:                      bar.go:  /* compiled into libbar.so */
       package "abc"                package "bar"
       import "bar"                 type empty struct { }
       ...                          var X empty
        foo(bar.X, ...)

In this case the fact that bar.X has zero size confuses the linker,
triggering bad behavior. Specifically when the a.out containing abc.go
is linked, you can wind up with the wrong flavor of dynamic relocation
applied to the text that refers to bar.X (link succeeds, but resulting
a.out is damaged). This incorrect relocation triggers either a warning
from the dynamic linker at runtime when the relocation is processed
(ex: "Symbol 'bar.X' causes overflow in <...> relocation") or an error
("while loading shared libraries: unexpected reloc type <XXX>").

To work around the linker problems, this patch looks for definitions
of global variables with zero size and converts the underlying type of
the variable to something with non-zero size (typically by adding a
dummy field to a structure or dummy element to an array). It then adds
typecasts/coercions to all of the uses of the variable. In effect this
makes the code in bar.go look like:

           bar.go:  /* compiled into libbar.so */
             package "bar"
             type empty struct { }
             type _tmp_generated struct { dummy uint8 }
             var X _tmp_generated
	     // exported type of "X" is set to 'empty', not '_tmp_generated'

For this to work, the bridge has to essentially rewrite any references
to such variables to insert type casts/conversions from the underlying
non-zero sized type to the high-level zero sized type.

Change-Id: Ic96b63fb0fff8e060f3aa4709d13edebdfd8ee89
Reviewed-on: https://go-review.googlesource.com/65990
Reviewed-by: Cherry Zhang <cherryyz@google.com>