gollvm: minor debugging fix
Switch up the order of a couple calls in the main driver routine to
insure that if --dump-ir is in effect, IR for the module gets dumped
before it is passed to the verifier. This makes it possible to inspect
the module IR in cases where verification does in fact fail.

Change-Id: Id89ea3d99494278e9acb84aca80b6e82b47e86cb
Reviewed-on: https://go-review.googlesource.com/65991
Reviewed-by: Cherry Zhang <cherryyz@google.com>