gollvm: support for module-scope optimizations
Improved/enhanced pass manager support, including inlining and loop
optimizations (although as of yet no LTO/PGO). The additional passes
can (for the moment) be turned off using the command line option
"--full-passes=0".

NB: this change causes a failure in the runtime/pprof package test

--- FAIL: TestCPUProfileInlining (...)
--- FAIL: TestCPUProfileMultithreaded (...)

This is a spurious failure; the LLVM optimizer is a bit too "smart"
and has deleted most of the contents of the "cpu-bound" function that
the test is trying to benchmark. I will follow up with a test fix.

Change-Id: I390e29eebeea0357fe931f2a2b952481863fd44a
Reviewed-on: https://go-review.googlesource.com/76023
Reviewed-by: Cherry Zhang <cherryyz@google.com>