gollvm: fix incorrect array size in debug meta-data
The code to compute array type size in bits for debug meta-data
was incorrect (it was using number of elements instead). Fix up
the recipe to use the correct type manager call. Interestingly,
this discrepancy between the array size and the array element
size doesn't get caught by the regular verifier, but once SROA
runs and the type is split into pieces, the verifier complains
about the insanity in the resulting fragment locations.

Change-Id: I26a227031883f7282eb19e273c16bc6aea396b21