gollvm: reduce compiler-generated const var duplication
At certain points during LLVM IR generation it is necessary to create
a variable to hold an anonymous LLVM constant (for example, if the
constant is an operand of a memcpy call). The helper method for
creating such vars was written to try to avoid creating duplicates
(same contents), but the commoning code was not doing the lookup
correctly. The lookup table key needs to be the constant value, not
the value associated with the resulting global variable (which
incorporates a name); this change fixes the duplicate check to use a
separate lookup table.

Change-Id: Iaa4075fd5d37889997d772f4bff1673d11cf01a8
Reviewed-on: https://go-review.googlesource.com/82175
Reviewed-by: Cherry Zhang <cherryyz@google.com>