gollvm: minor cmake cleanup
Rework the cmake rules that handle the GMP/MPC/MPFR dependencies
so as to eliminate the need for a "prebuild" step when building
gollvm (you can now just do cmake then ninja). Remove the obsolete
'Parser' unit test directory; this was originally intended to
have separate tests of frontend machinery, but this is not practical
(and the test themselves were incomplete).

Change-Id: I22af6cdc950e693cfb3fe71411e0364d41030894