gollvm: reorganization/refactoring in preparation for libgo cmake
Reorganize the gollvm source code in preparation for incorporating
cmake rules for the libgo build. The subdirectory that contained the
bridge code has been renamed from 'llvm-gofrontend' to 'bridge', and
'gofrontend' itself is now a child of 'gollvm' and not the bridge.
Minor updates to the README.md to track these changes.

Change-Id: I1ac3441160e680e6c09972e754ee18180f4b2701