gollvm: sync up with recent gofrontend enhancements
This patch syncs up the gollvm bridge code with recent changes
in gofrontend, including

 * new "decl_var" parameter in Backend::local_variable method
 * debug_escape_hash Gogo parameter

For the LLVM bridge version of If Backend::local_variable, if a
non-NULL decl var is provided, we check to make sure that the variable
is part of the same function, then arrange for the new variable to
share the old variable's alloca instruction (which should give us the
desired behavior). NB: at some point there may be additional work
needed when lifetime annotations are added for all local vars.

Change-Id: I8e9825968ffb16d8ed1700f0395b44992834c9f0