gollvm: sync up with recent gofrontend enhancements
This patch syncs up the gollvm bridge code with recent changes
in gofrontend, including

 * new "decl_var" parameter in Backend::local_variable method
 * debug_escape_hash Gogo parameter

For the LLVM bridge version of If Backend::local_variable, if a
non-NULL decl var is provided, we check to make sure that the variable
is part of the same function, then arrange for the new variable to
share the old variable's alloca instruction (which should give us the
desired behavior). This patch also changes lifetime start/end
annotation insertion avoid adding lifetimes for a declvar variable,
since its extent is based not on its enclosing block but on the var
whose alloca it shares.

Change-Id: I8e9825968ffb16d8ed1700f0395b44992834c9f0