gollvm: include auto-generated Go files in Go stdlib packages
This adds in machinery to create the various auto-generated Go files
that are part the Go std library. With this patch you can build
packages such as "syscall".  For one specific file (gen-sysinfo.go)
we're using an archived copy, since the machinery to generate it is
not quite ready (will be submitted in a subsequent patch).

Change-Id: I1970c26b7198f00b5ef277314a197afb6f95f733