gollvm: work around multiple calls of set_placeholder_pointer_type
The frontend may call set_placeholder_pointer_type multiple times
with the same type. After the first call, the type is marked as
non-placeholder, and the types referencing this type are finalized.
At the second call, if this type is updated, but the referencing
types are not, there will be inconsistency, and will result in
type mismatch failures later.

One example is pointer to circular function type. If we have a
placeholder pointer type PPT*, pointing to a circular function
type CFT*, the first call of set_placeholder_pointer_type resolves
PPT* to CFT**. Later set_placeholder_pointer_type is called the
second time with CFT replaced by the actual function. If we change
the Btype's underlying type from CFT** to function**, other type,
e.g. struct { CFT** } will not be updated because CFT** is not a
placeholder, causing inconsistency.

This fixes another assertion failure when building libgo with
escape analysis on.

Change-Id: I767231c77ef7225fb5305386a205a906b688c60c