gollvm: introduce llvm-godumpspec tool for libgo build
Introduce a new tool 'llvm-godumpspec', which performs a role similar
to GCC's "-fdump-go-spec" command line option within the libgo build.
In this case rather than walking the debug meta-data during the
compile of a given C file, this tool walks the emitted DWARF instead,
generating Go versions of various C constants, types, macros, and
variables.

This version is only partially functional; there is some support for
types and some support for variables. Things yet to be dealt with:

  - macros
  - reliable generation of unused types

NB: this monolithic patch needs to be split into smaller chunks
for review.

Notes:
- needs work for the clang case to handle forward struct type
  reference. Might make sense to switch to a 2-pass scheme in
  which all sizes and alignments are handled on the first pass,
  emit comes on the second pass.

Change-Id: Ic138b5523a88203258c697c1861e6b910b450d20