gollvm: clean up of external dependency build rules
Tweak the cmake rules for libgmp/libmpfr/libmpc to insure that the
build works on systems that do not have any GMP/MPC/MPFR bits
installed (headers, etc). Update the dependencies to insure that the
external dep build completes before we start building the bridge.

This change based in part by a patch by Serban Simu.

Change-Id: I76012e6425a3aa512ded6758050b0daf46a414fd