gollvm: cmake rules for the gotools test targets
Support for testing the go tools, including 'go', 'vet', and 'cgo', by
creating a small GOPATH-like environment and running the package tests
for them via "go test". This is part 1 of 2 of the gotools check targets.

Note: a set of testpoints in the cmd/go test fail currently due to the
use of filenames with embedded spaces (which don't play well with the
wrapper):

FAIL: TestBadImportsEasy
FAIL: TestBadImportsEasySub
PASS: TestBadImportsGoInstallShouldFail
FAIL: TestBadImportsHard

These failures should go away once we do away with the wrapper entirely.

Change-Id: I206450911b5ce2c5de47061f5f553b00cf35dd1f