gollvm: support for passing assembly source files to driver
Support compilations in which input files are assembly source code and
not Go source code (in which case the driver simply invokes the
assembler on them). Do not permit mixing of assembly and source inputs
(to be compatible with gccgo). No support yet for *.S files (which
need to be run through the preprocessor).

Change-Id: I7ea20c96a2a16920cd7b75fdd6aa8c96c24037df